% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perror.R
\name{perror}
\alias{perror}
\title{Empirical cumulative distribution functions of conditional errors}
\usage{
perror(p, xs)
}
\arguments{
\item{p}{Vector of quantiles}

\item{xs}{Indices of the test observations for which the conditional error
CDFs are desired. Defaults to all test observations given in the call of
\code{quantForestError}.}
}
\value{
If either \code{p} or \code{xs} has length one, then a vector is
  returned with the desired probabilities. If both have length greater than
  one, then a \code{data.frame} is returned, with rows corresponding to the
  inputted \code{xs} and columns corresponding to the inputted \code{p}.
}
\description{
Returns probabilities from the estimated cumulative distribution function
of the conditional error distribution associated with each test prediction.
}
\examples{
\donttest{
# get the probability that the error associated with each test prediction
# is less than -5 and the probability that the error associated with each
# test prediction is less than 5
perror(c(-5, 5))

# same as above but only for the first three test observations
perror(c(-5, 5), 1:3)
}
}
\seealso{
\code{\link{quantForestError}}
}
\author{
Benjamin Lu \code{<b.lu@berkeley.edu>}; Johanna Hardin \code{<jo.hardin@pomona.edu>}
}
