% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qerror.R
\name{qerror}
\alias{qerror}
\title{Empirical quantile functions of conditional errors}
\usage{
qerror(q, xs)
}
\arguments{
\item{q}{Vector of probabilities}

\item{xs}{Indices of the test observations for which the conditional error
CDFs are desired. Defaults to all test observations given in the call of
\code{quantForestError}.}
}
\value{
If either \code{q} or \code{xs} has length one, then a vector is
  returned with the desired probabilities. If both have length greater than
  one, then a \code{data.frame} is returned, with rows corresponding to the
  inputted \code{xs} and columns corresponding to the inputted \code{q}.
}
\description{
Returns quantiles from the estimated quantile function of the conditional
error distribution associated with each test prediction.
}
\examples{
\donttest{
# get the 0.25 and 0.8 quantiles of the error distribution associated
# with each test observation
qerror(c(0.25, 0.8))

# same as above but only for the first three test observations
qerror(c(0.25, 0.8), 1:3)
}
}
\seealso{
\code{\link{quantForestError}}
}
\author{
Benjamin Lu \code{<b.lu@berkeley.edu>}; Johanna Hardin \code{<jo.hardin@pomona.edu>}
}
