% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perror.R
\name{perror}
\alias{perror}
\title{Estimated conditional prediction error CDFs}
\usage{
perror(q, xs)
}
\arguments{
\item{q}{A vector of quantiles.}

\item{xs}{A vector of the indices of the test observations for which the
conditional error CDFs are desired. Defaults to all test observations
given in the call of \code{quantForestError}.}
}
\value{
If either \code{q} or \code{xs} has length one, then a vector is
  returned with the desired probabilities. If both have length greater than
  one, then a \code{data.frame} of probabilities is returned, with rows
  corresponding to the inputted \code{xs} and columns corresponding to the
  inputted \code{q}.
}
\description{
Returns probabilities from the estimated conditional cumulative distribution
function of the prediction error associated with each test observation.
}
\examples{
\donttest{
# get the probability that the error associated with each test
# prediction is less than -4 and the probability that the error
# associated with each test prediction is less than 7
perror(c(-4, 7))

# same as above but only for the first three test observations
perror(c(-4, 7), 1:3)
}
}
\seealso{
\code{\link{quantForestError}}
}
\author{
Benjamin Lu \code{<b.lu@berkeley.edu>}; Johanna Hardin \code{<jo.hardin@pomona.edu>}
}
