% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qerror.R
\name{qerror}
\alias{qerror}
\title{Estimated conditional prediction error quantile functions}
\usage{
qerror(p, xs)
}
\arguments{
\item{p}{A vector of probabilities.}

\item{xs}{A vector of the indices of the test observations for which the
conditional error quantiles are desired. Defaults to all test observations
given in the call of \code{quantForestError}.}
}
\value{
If either \code{p} or \code{xs} has length one, then a vector is
  returned with the desired quantiles. If both have length greater than
  one, then a \code{data.frame} of quantiles is returned, with rows
  corresponding to the inputted \code{xs} and columns corresponding to the
  inputted \code{p}.
}
\description{
Returns quantiles of the estimated conditional error distribution associated
with each test prediction.
}
\examples{
\donttest{
# get the 0.25 and 0.8 quantiles of the error distribution
# associated with each test observation
qerror(c(0.25, 0.8))

# same as above but only for the first three test observations
qerror(c(0.25, 0.8), 1:3)
}
}
\seealso{
\code{\link{quantForestError}}
}
\author{
Benjamin Lu \code{<b.lu@berkeley.edu>}; Johanna Hardin \code{<jo.hardin@pomona.edu>}
}
